+++ src/django_otp/plugins/otp_static/tests.py
@@ -1,7 +1,8 @@
 from django.db import IntegrityError
+from django.test.utils import override_settings
 
 from django_otp.forms import OTPAuthenticationForm
-from django_otp.tests import TestCase
+from django_otp.tests import TestCase, ThrottlingTestMixin
 
 from .lib import add_static_token
 from .models import StaticDevice
@@ -161,3 +162,25 @@ class AuthFormTest(TestCase):
         alice = form.get_user()
         self.assertEqual(alice.get_username(), 'alice')
         self.assertIsNotNone(alice.otp_device)
+
+
+@override_settings(
+    OTP_STATIC_THROTTLE_FACTOR=1,
+)
+class ThrottlingTestCase(ThrottlingTestMixin, TestCase):
+    def setUp(self):
+        try:
+            user = self.create_user('alice', 'password')
+        except IntegrityError:
+            self.skipTest("Unable to create a test user.")
+        else:
+            self.device = user.staticdevice_set.create()
+            self.device.token_set.create(token='valid1')
+            self.device.token_set.create(token='valid2')
+            self.device.token_set.create(token='valid3')
+
+    def valid_token(self):
+        return self.device.token_set.first().token
+
+    def invalid_token(self):
+        return 'bogus'